requirements.txt
numpy
seaborn
sklearn
Pillow
